medical/ai-voice-response-system/unified-platform/requirements.txt
# 完全統合AI音声応答プラットフォーム - 依存パッケージ

# コアフレームワーク
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# 音声認識
openai-whisper>=20231117

# 音声合成
pyttsx3>=2.90
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0

# AI応答生成
openai>=1.3.0

# HTTPクライアント（法務システム統合用）
httpx>=0.25.0

# その他
websockets>=12.0
